client::send: fix warning

+++ src/wsclient.cpp
@@ -735,7 +735,7 @@ namespace ws {
 
 			auto& strm = impl->zstrm_out.value();
 
-			strm.avail_in = payload.size();
+			strm.avail_in = (int)payload.size();
 			strm.next_in = (uint8_t*)payload.data();
 
 			do {
